requirements.txt
APScheduler==3.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
cryptography==3.0
decorator==4.4.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pyaes==1.6.1
pycparser==2.20
Pyrogram==0.18.0
PySocks==1.7.1
python-dotenv==0.14.0
python-telegram-bot==12.8
pytz==2020.1
six==1.15.0
tornado==6.0.4
tzlocal==2.1
Werkzeug==1.0.1
emoji
TgCrypto
jdatetime
